peformance better ,but lose accuracy

+++ gbfloat_fast.c
@@ -55,7 +55,7 @@ float *get_pixel(Image img, int x, int y) {
 
 float gd(float a, float b, float x) {
     float c = (x - b) / a;
-    return exp((-.5) * c * c) / (a * sqrt(2 * PI));
+    return expf((-.5) * c * c) / (a * sqrt(2 * PI));
 }
 
 FVec make_gv(float a, float x0, float x1, unsigned int length, unsigned int min_length) {
@@ -102,8 +102,8 @@ Image gb_h(Image a, FVec gv) {
     for (y = 0; y < a.dimY; y++) {
         for (x = 0; x < a.dimX; x++) {
             pc = get_pixel(b, x, y);
-            deta = fmin(fmin(fmin(a.dimY - y - 1, y), fmin(a.dimX - x - 1, x)), gv.min_deta);
-            double sum1 = 0, sum2 = 0, sum3 = 0;
+            deta = fminf(fminf(fminf(a.dimY - y - 1, y), fminf(a.dimX - x - 1, x)), gv.min_deta);
+            float sum1 = 0, sum2 = 0, sum3 = 0;
             for (i = deta; i < gv.length - deta; ++i) {
                 sum1 += gv.data[i] * get_pixel(a, x - ext + i, y)[0];
                 sum2 += gv.data[i] * get_pixel(a, x - ext + i, y)[1];
@@ -129,8 +129,8 @@ Image gb_v(Image a, FVec gv) {
     for (y = 0; y < a.dimY; y++) {
         for (x = 0; x < a.dimX; x++) {
             pc = get_pixel(b, x, y);
-            deta = fmin(fmin(fmin(a.dimY - y - 1, y), fmin(a.dimX - x - 1, x)), gv.min_deta);
-            double sum1 = 0, sum2 = 0, sum3 = 0;
+            deta = fminf(fminf(fminf(a.dimY - y - 1, y), fminf(a.dimX - x - 1, x)), gv.min_deta);
+            float sum1 = 0, sum2 = 0, sum3 = 0;
             for (i = deta; i < gv.length - deta; ++i) {
                 sum1 += gv.data[i] * get_pixel(a, x, y - ext + i)[0];
                 sum2 += gv.data[i] * get_pixel(a, x, y - ext + i)[1];
@@ -208,5 +208,5 @@ int main(int argc, char **argv) {
 
 //begin to try something
 
-//can l change all doubles to float
-//do some minus optimization (change all exp to exp and fmin to fmin)
+//can l change all double to float
+//do some minus optimization (change all exp to exp and fmin to fminf)
